Split the Ligerito channels by concern, and make the reductions methods
Both `channel.rs` files held three concerns each: the oracle handle, the
queued relation, and the channel itself, plus two free functions doing
the actual work. Each is now its own file under a `channel/` module.

`prove_ligerito` and `verify_ligerito` were free functions taking their
subject first, so they are now `LigeritoProverChannel::prove` and
`LigeritoVerifierChannel::verify`, reached from `finish`.

`batch_relations` becomes `QueuedRelation::batch`, and it now takes the
already-sampled `lambda` rather than the channel. That moves the
security-critical ordering — draw the coefficient only once every claim
it combines is bound to the transcript — to the call site, where it is
visible, instead of hiding it inside a helper.

Both batching docs now name the deviation from the paper: [NA25] section
4.3 batches with `k` independent coefficients, where powers of one cost a
single challenge and trade `1 / |F|` of error for `(k - 1) / |F|`.

Both compilers' fields gained the header docs they were missing.
